% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrix.R
\name{confusao}
\alias{confusao}
\title{Plota matriz de confusao, versão 2}
\usage{
confusao(xtab)
}
\arguments{
\item{xtab}{Tabela cruzamento de dados observados versus dados preditos.}
}
\value{
Um plot.
}
\description{
Plota matriz de confusao, versão 2
}
\author{
Ricardo O. Perdiz, \email{ricoperdiz@gmail.com}
}
