% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrix.R
\name{confusao_lab}
\alias{confusao_lab}
\title{Plota matriz de confusão, estilo LABOTAM}
\usage{
confusao_lab(xtab, add_CP = FALSE)
}
\arguments{
\item{add_CP}{Adiciona a taxa de predições corretas? Padrão é não adicionar.}
}
\description{
Plota matriz de confusão, estilo LABOTAM
}
