% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nirdf.R
\name{is_nirdf}
\alias{is_nirdf}
\title{Testa se o objeto é um nirdf}
\usage{
is_nirdf(x)
}
\arguments{
\item{x}{Um objeto.}
}
\value{
`TRUE` se o objeto herda da classe `nirdf`
}
\description{
Esta função retorna `TRUE` para nirdf, e `FALSE` para outros objetos que não se adequam ao conceito de um nirdf.
}
