% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{lda_lab}
\alias{lda_lab}
\title{Executa LDA 70-30 por padrão}
\usage{
lda_lab(df, nir_params_path, aleatorizacao, teste_n = 0.3)
}
\arguments{
\item{df}{Um dataframe}

\item{nir_params_path}{Um arquivo de parâmetros}
}
\value{

}
\description{
Executa LDA 70-30 por padrão
}
\author{
Alberto Vicentini (INPA), \email{alberto.vicentini@inpa.gov.br}
}
