% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{lda_lab_batch}
\alias{lda_lab_batch}
\title{Executa LDA 70-30 padrão em um for loop. Apropriado para listas de data.frames.}
\usage{
lda_lab_batch(df, aleatorizacao, teste_n = NULL)
}
\arguments{
\item{df}{Um dataframe.}
}
\value{

}
\description{
Executa LDA 70-30 padrão em um for loop. Apropriado para listas de data.frames.
}
\author{
Ricardo de Oliveira Perdiz, \email{ricardoperdiz@yahoo.com}; Alberto Vicentini (INPA), \email{alberto.vicentini@inpa.gov.br}
}
