% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{lda_lab_plot}
\alias{lda_lab_plot}
\title{Plota resultado de uma LDA LOO.}
\usage{
lda_lab_plot(train_set, train_model, train_spp, col = FALSE, color_p = "")
}
\arguments{
\item{train_model}{}
}
\value{

}
\description{
Copiado desta [postagem](https://www.r-bloggers.com/linear-quadratic-and-regularized-discriminant-analysis/).
}
