% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_loo.R
\name{lda_loo}
\alias{lda_loo}
\title{LDA Leave One Out}
\usage{
lda_loo(df, nir_params_path, save.csv = FALSE)
}
\arguments{
\item{nir_params_path}{A path that points to a NIR parameter file.}

\item{save.csv}{}
}
\value{

}
\description{
LDA Leave One Out
}
