% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{lda_loo_batch}
\alias{lda_loo_batch}
\title{Faz LDA LOO em batch}
\usage{
lda_loo_batch(df, times = 1, dataset_name = NULL, wd = wd)
}
\description{
Faz LDA LOO em batch
}
\author{
Ricardo de Oliveira Perdiz, \email{ricardoperdiz@yahoo.com}; Alberto Vicentini (INPA), \email{alberto.vicentini@inpa.gov.br}
}
