% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nir_dad}
\alias{nir_dad}
\title{Conjunto de dados NIR processado e convertido em objeto nirdf}
\format{
Um data.frame de `r nrows(nir_data)` linhas e `r ncol(nir_data)` colunas.
\describe{
\item{especimenid}{Identificador da amostras.}
\item{SP1}{Nome da espécie da amostras.}
\item{face}{De qual face da folha a leitura foi obtida? Abaxial (inferior) ou adaxial (superior)?}
\item{coletor}{Nome do coletor da amostra.}
\item{number}{Número da amostra coletada.}
\item{Colunas que começam com X}{Cada uma corresponde a uma faixa do espectro, e possuem valores de absorbância.}
}
}
\usage{
nir_dad
}
\description{
Um subconjunto de dados oriundos da tese do autor ([Perdiz, 2020](https://repositorio.inpa.gov.br/handle/1/36948)), que consiste em uma série de leituras do espectro infravermelho de folhas de um grupo de árvores de distribuição neotropical. Dados convertidos para um objeto nirdf com a função `nirdf()`.
}
\keyword{datasets}
