% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-nir_data.R
\docType{data}
\name{nir_data}
\alias{nir_data}
\title{Dataset of Near Infra Red data}
\format{A dataframe with 541 rows and 1561 columns
\describe{
\item{especimenid}{Specimen identification}
\item{SP1}{Species id}
\item{read}{Read number id}
\item{face}{Which leaf surface?}
\item{Columns that starts with X}{each one of these is a spectrum}
}}
\usage{
nir_data
}
\description{
A dataset containing a subset from the author's thesis, that consists of reads of  near infrared spectrum of leaves of a group of neotropical plants
}
\keyword{datasets}
