% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nirdad}
\alias{nirdad}
\title{Conjunto de dados NIR processado e convertido em objeto nirdf}
\format{
Um data.frame de `r nrows(nir_data)` linhas e `r ncol(nir_data)` colunas.
\describe{
\item{especimenid}{Identificador da amostras.}
\item{SP1}{Nome da espécie da amostras.}
\item{face}{De qual face da folha a leitura foi obtida? Abaxial (inferior) ou adaxial (superior)?}
\item{coletor}{Nome do coletor da amostra.}
\item{number}{Número da amostra coletada.}
\item{Colunas que começam com X}{Cada uma corresponde a uma faixa do espectro, e possuem valores de absorbância.}
}
}
\usage{
nirdad
}
\description{
Um subconjunto de dados oriundos da tese do autor, que consiste em uma série de leituras do espectro infravermelho de folhas de um grupo de árvores de distribuição neotropical.
}
\keyword{datasets}
