% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nirdf.R
\name{nirdf}
\alias{nirdf}
\title{Cria um objeto nirdf}
\usage{
nirdf(x, category, measure_columns, measure_columns_prefix = NULL)
}
\arguments{
\item{x}{Um data.frame, ou um data.table.}

\item{category}{Um nome apontando o nome da variável categórica dentro do argument x.}

\item{measure_columns}{Um vetor contendo nomes de coluna que contêm as variáveis NIR.}
}
\value{
Um objeto nirdf.
}
\description{
Cria um objeto nirdf
}
