% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{pegaespectro}
\alias{pegaespectro}
\title{Cria uma funcao que seleciona 1 spectro aleatorio para cada uind}
\usage{
pegaespectro(x, tb, id)
}
\value{

}
\description{
Cria uma funcao que seleciona 1 spectro aleatorio para cada uind
}
\author{
Alberto Vicentini (INPA), \email{alberto.vicentini@inpa.gov.br}
}
