% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{pegateste}
\alias{pegateste}
\title{Cria uma funcao que seleciona 1 spectro aleatorio para cada uind}
\usage{
pegateste(x, dd, ind, teste_n = 0.3)
}
\description{
Cria uma funcao que seleciona 1 spectro aleatorio para cada uind
}
\author{
Alberto Vicentini (INPA), \email{alberto.vicentini@inpa.gov.br}
}
