% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.nirdf}
\alias{plot.nirdf}
\title{Plot a nirdf object.}
\usage{
\method{plot}{nirdf}(
  x,
  category,
  remove_cols = NULL,
  xlabel = parse(text = "Wavenumber (cm^-1)"),
  ylabel = "Absorbance",
  legend_position = "topright",
  color = NULL,
  ...
)
}
\arguments{
\item{x}{A nirdf object.}

\item{category}{A string pointing the name of the category variable of nirdf object.}

\item{remove_cols}{Character vector pointing name of columns to be removed.}

\item{xlabel}{X axis label.}

\item{ylabel}{Y axis label.}

\item{legend_position}{Legend position, as done in function legend.}

\item{color}{Vector of colors. If not provided, function will provide one. Attention to the length of this vector. It should be the same length as the number of categories.}

\item{...}{}
}
\value{
A plot
}
\description{
\lifecycle{experimental}
}
\examples{
str(nirdf)
# nirdf object is a data.frame containing a column for categories, `SP1`, an id column, `especimenid`, and all the rest of columns containg Near Infrared data.
plot(nirdf, category = "SP1", remove_cols = "especimenid")
}
