% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dirNIRraw.R
\name{read_dirNIRraw}
\alias{read_dirNIRraw}
\title{Lê dados NIR em formato bruto obtido em um espectrofotômetro Thermo Nicollet \(Thermo Fischer Scientific, Waltham, Massachusetts, USA\), modelo Antaris II FT-NIR Analyzer, hospedado no herbário do INPA, Amazonas, Brasil}
\usage{
read_dirNIRraw(arq, add_nir_id = TRUE)
}
\arguments{
\item{arq}{Vetor de texto contendo caminho para um arquivo contendo dados de infravermelho próximo em formato bruto. Geralmente o dado será um arquivo de 1557 linhas, em que a primeira coluna conterá o número identificador do espectro, e a segunda coluna conterá o valor de absorbância NIR.}

\item{add_nir_id}{Vetor lógico, `TRUE` ou `FALSE`, para saber se deve ser adicionado a letra `X` antes de cada variável NIR no arquivo final}
}
\value{
Um data.frame em formato amplo (`wide`  em inglês) em que cada linha conterá as 1557 colunas de dados NIR..
}
\description{
Lê dados NIR em formato bruto obtido em um espectrofotômetro Thermo Nicollet \(Thermo Fischer Scientific, Waltham, Massachusetts, USA\), modelo Antaris II FT-NIR Analyzer, hospedado no herbário do INPA, Amazonas, Brasil
}
\examples{
# read file with raw data
nir_raw <- system.file("extdata", "nir_raw.csv", package = "NIRtools")
nir_raw <- read_dirNIRraw(arq = nir_raw, add_nir_id = TRUE)
# check first rows and 5 first columns of data
dim(nir_raw)
names(nir_raw)
head(nir_raw)[,1:10]
}
