% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_dirNIRraw_alt}
\alias{read_dirNIRraw_alt}
\title{Title}
\usage{
read_dirNIRraw_alt(
  list_files,
  source = c("nano", "inno"),
  wavelength_var_name_partial = "Wavelength",
  absorbance_var_name_partial = "Absorbance"
)
}
\arguments{
\item{absorbance_var_name_partial}{}
}
\value{
Um data.frame
}
\description{
Title
}
