% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_dlp_nano}
\alias{read_dlp_nano}
\title{Importa arquivo da Texas Instruments}
\usage{
read_dlp_nano(file = "", sep = "\\n")
}
\arguments{
\item{file}{Vetor de texto indicando caminho a um arquivo com dados de leitura espectroscópica obtidos com um aparelho Texas Instruments, modelo DLPNIRNANOEVM.}

\item{sep}{Separador}
}
\value{
Um data.frame
}
\description{
Importa arquivo da Texas Instruments
}
\examples{
arq <- system.file("extdata", "nir_raw3.csv", package = "NIRtools")
read_dlp_nano(file = arq)
}
