% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_inno}
\alias{read_inno}
\title{Importa arquivo da INNO, modelo NIR-S-G1}
\usage{
read_inno(file = "", sep = "\\n")
}
\arguments{
\item{file}{Vetor de texto indicando caminho a um arquivo com dados de leitura espectroscópica obtidos com um aparelho INNO, modelo NIR-S-G1.}

\item{sep}{Separador}
}
\value{
Um data.frame
}
\description{
Importa arquivo da INNO, modelo NIR-S-G1
}
\examples{
arq <- system.file("extdata", "nir_raw4.csv", package = "NIRtools")
read_inno(file = arq)
}
