% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_NIRraw}
\alias{read_NIRraw}
\title{Read near infrared (NIR) spectroscopy data in raw format}
\usage{
read_NIRraw(file, add_nir_id = TRUE)
}
\arguments{
\item{file}{A path to a file containing near infrared (NIR) spectroscopy raw data. This means that it expects to be a file with 1557 rows, in which first column contains identification of spectrum number, and second column the NIR absorbance.}
}
\value{
A data.frame in `wide` format in which each row contains 1557 columns of near infrared spectroscopy data.
}
\description{
Read near infrared (NIR) spectroscopy data in raw format
}
\examples{

# read file with raw data
nir_raw <- read_NIRraw("data/nir_raw.csv")
# check first rows and 5 first columns of data
dim(nir_raw)
names(nir_raw)
head(nir_raw)[, 1:10]
}
