% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{resumo_lda}
\alias{resumo_lda}
\title{Gera um resumo dos dados da LDA 70-30 a partir do resultado gerado pela função beto_lda7.}
\usage{
resumo_lda(resultado.bruto, acertosporpermutacao)
}
\description{
Gera um resumo dos dados da LDA 70-30 a partir do resultado gerado pela função beto_lda7.
}
\author{
Alberto Vicentini (INPA), \email{alberto.vicentini@inpa.gov.br}
}
