% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{save_lda}
\alias{save_lda}
\title{Salva o resultado da LDA 70-30 (padrão) na forma de um pdf com histograma, e o conjunto de dados em um RDS.}
\usage{
save_lda(lda_lab_list, save.pdf = TRUE, save.RDS = TRUE, wd = NULL)
}
\arguments{
\item{save.pdf}{Salva pdf?}

\item{save.RDS}{Salva RDS?}

\item{wd}{Diretório de trabalho}
}
\value{

}
\description{
Salva o resultado da LDA 70-30 (padrão) na forma de um pdf com histograma, e o conjunto de dados em um RDS.
}
\author{
Ricardo de Oliveira Perdiz, \email{ricardoperdiz@yahoo.com}; Alberto Vicentini (INPA), \email{alberto.vicentini@inpa.gov.br}
}
