% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{save_NIRAresults}
\alias{save_NIRAresults}
\title{Salva o resultado das análises NIR (NIRA)}
\usage{
save_NIRAresults(file_output = "", run_NIRA = FALSE)
}
\value{

}
\description{
Salva o resultado das análises NIR (NIRA)
}
\author{
Ricardo de Oliveira Perdiz, \email{ricardoperdiz@yahoo.com}
}
