% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_NIRdataset}
\alias{summarise_NIRdataset}
\title{Summarise NIR datasets}
\usage{
summarise_NIRdataset(dframe, params_file_path)
}
\arguments{
\item{dframe}{Raw NIR data.}

\item{params_file_path}{Location of NIR parameter file.}
}
\description{
Summarise NIR datasets
}
