% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{visualiza_res_modelos}
\alias{visualiza_res_modelos}
\title{Tabela com resultados de acertos por espécie da LDA}
\usage{
visualiza_res_modelos(resultado_bruto)
}
\arguments{
\item{resultado_bruto}{}
}
\description{
Tabela com resultados de acertos por espécie da LDA
}
