% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_NIRparams}
\alias{write_NIRparams}
\title{Write NIR parameters file to disk}
\usage{
write_NIRparams(
  file = "",
  wd = ".",
  surface = "",
  reads = "",
  nir_variables = "",
  subset_file = "",
  individual_id = "",
  individual_list = NULL,
  surface_id = "",
  group_id = "",
  nir_id = "X"
)
}
\arguments{
\item{file}{Dataset name.}

\item{wd}{Location of working directory. If not provided, current working directory will be used one.}

\item{surface}{Which leaf surface (abaxial, adaxial, both).}

\item{reads}{Do all the reads will be taken into account (all, mean)?}

\item{nir_variables}{Which NIR variables will be used (all, subset).}

\item{subset_file}{Location of NIR variable subset file, in case option selected in `nir_variables` was `subset`.}

\item{individual_id}{Which variable name corresponds to the individual?}

\item{individual_list}{Location of a file containing list of specimens to subset data when building datasets with `build_NIRdataset` function.}

\item{surface_id}{Which variable name corresponds to the leaf surface?}

\item{group_id}{Which variable name corresponds to the group category?}

\item{nir_id}{A string that can be used to grep column names containing NIR data. Default value is '`X`, which precedes all columns with NIR data.}
}
\value{
A message that indicates the file has been saved to disk.
}
\description{
Write NIR parameters file to disk
}
\examples{
\dontrun{
write_NIRparams(file = "teste", wd = ".",
 reads = "mean", surface = "abaxial",
  nir_variables = "all", surface_id = "face",
   individual_id = "especimenid",
   individual_list = NULL, group_id = "SP1",
    nir_id = "X")
read_NIRparams("teste-NIR.txt")
readLines("teste-NIR.txt")
}
}
