% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nir_raw}
\alias{nir_raw}
\title{Conjunto de dados NIR em formato bruto}
\format{
Um dataframe com 1557 linhas e 2 colunas
\describe{
\item{wavenumber}{Número da onda}
\item{absorbance}{Absorbância}
}
}
\usage{
nir_raw
}
\description{
Subconjunto de dados utilizado na tese do autor do pacote ([Perdiz, 2020](https://repositorio.inpa.gov.br/handle/1/36948)), consistindo em leituras espectrais do infravermelho próximo de um grupo de plantas de distribuição Neotropical. Formato dos dados consiste em uma observação de leitura espectral por linha.
}
\details{
`r (nir_raw)`
}
\keyword{datasets}
