% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.nirdf}
\alias{plot.nirdf}
\title{Plota um objeto nirdf.}
\usage{
\method{plot}{nirdf}(
  x,
  category,
  remove_cols = NULL,
  xlabel = parse(text = "Wavenumber (cm^-1)"),
  ylabel = "Absorbance",
  legend_position = "topright",
  cex_pt = 0.05,
  cex_leg = 0.5,
  color = NULL,
  ...
)
}
\arguments{
\item{x}{Um objeto nirdf.}

\item{category}{Nome da coluna contendo a variável categórica do objeto nirdf.}

\item{remove_cols}{Vetor de texto contendo os nomes das colunas a serem removidos.}

\item{xlabel}{Texto do eixo X.}

\item{ylabel}{Texto do eixo Y.}

\item{legend_position}{Posição da legenda, como adotado na função `legend()`.}

\item{cex_pt}{Tamanho adotado no argumento cex da função `points()`.}

\item{cex_leg}{Tamanho adotado no argumento cex da função `legend()`.}

\item{color}{Vetor de cores. Se não fornecido, a função providenciará uma seleção automaticamente. Atenção ao comprimento deste vetor. Ele deve ser de tamanho mínimo ao número de categorias do objeto nirdf.}

\item{...}{Funções adicionais a serem fornecidas à função `plot()`.}
}
\value{
Um plot.
}
\description{
\lifecycle{experimental}
}
\examples{
# Load nir data
library("NIRtools")
data(nir_data)
nirdad <- nirdf(nir_data, "SP1",
 measure_columns = grep("^X", names(nir_data), value = TRUE),
  measure_columns_prefix = "X")
plot(nirdad, "SP1")
}
