% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrix.R
\name{plotamatriz}
\alias{plotamatriz}
\title{Plota matriz de confusão}
\usage{
plotamatriz(
  matriz,
  bg.cols = NULL,
  txt.cols = NULL,
  valcex = 1,
  cexaxis = 1,
  lasx = lasx,
  lasy = lasy
)
}
\arguments{
\item{matriz}{matrix com colnames e rownames}

\item{bg.cols}{matrix com valores de cores para background das células}

\item{txt.cols}{matrix com valores de cores para o texto dos valores das células}

\item{valcex}{cex do texto dos valores das células}

\item{cexaxis}{cex do texto dos eixos = nomes de colunas e linhas}

\item{lasx}{Orientação do texto do eixo x.}

\item{lasy}{Orientação do texto do eixo y.}
}
\value{
Um plot.


}
\description{
Plota matriz de confusão
}
\author{
Alberto Vicentini, \email{vicentini.beto@gmail.com}
}
