% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_NIRparams}
\alias{read_NIRparams}
\title{Lê metadados de um subconjunto de dados NIR a partir de um arquivo de texto}
\usage{
read_NIRparams(arq)
}
\arguments{
\item{arq}{Vetor de texto indicando caminho a um arquivo.}
}
\value{
Um data.frame
}
\description{
Lê metadados de um subconjunto de dados NIR a partir de um arquivo de texto
}
\examples{
params_file_path <- system.file("extdata", "conjunto1-NIRparams.txt", package = "NIRtools")
read_NIRparams(arq = params_file_path)
}
