% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_NIRparams}
\alias{read_NIRparams}
\title{Read NIR parameters file from disk}
\usage{
read_NIRparams(file)
}
\arguments{
\item{file}{A path to a file.}
}
\value{
A data.frame
}
\description{
Read NIR parameters file from disk
}
\examples{
# Input file on vignette folder
read_NIRparams("test-NIR.txt")
}
