% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nir_pkg.R
\name{read_NIRparams}
\alias{read_NIRparams}
\title{Read NIR parameters file from disk}
\usage{
read_NIRparams(file)
}
\arguments{
\item{file}{A path to a file.}
}
\description{
Read NIR parameters file from disk
}
