% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{run_NIRA}
\alias{run_NIRA}
\title{Run Near Infra Red Analysis (NIRA) and plot results}
\usage{
run_NIRA(dataset, nir_params_path, outfig = ".",
  run_analysis = c("plot_raw", "PCA", "PCA_plot", "LDA", "LDA_plot",
  "LOO", "LOO_plot"), aleatorizacao = 200, save.csv = TRUE,
  add_CP = FALSE, use_date = FALSE, col_p = cores, use_col = TRUE,
  outgroup = "")
}
\description{
Run Near Infra Red Analysis (NIRA) and plot results
}
