% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{save_NIRAresults}
\alias{save_NIRAresults}
\title{Save Near Infra Red Analysis (NIRA) results}
\usage{
save_NIRAresults(file_output = "", run_NIRA = FALSE)
}
\value{

}
\description{
Save Near Infra Red Analysis (NIRA) results
}
\author{
Ricardo de Oliveira Perdiz, \email{ricardoperdiz@yahoo.com}
}
