% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_NIRdataset}
\alias{summarise_NIRdataset}
\title{Sumariza conjuntos de dados NIR.}
\usage{
summarise_NIRdataset(dframe, params_file_path, use_cwd = TRUE)
}
\arguments{
\item{dframe}{Dados brutos NIR.}

\item{params_file_path}{Localização do arquivo de parâmetros NIR.}

\item{use_cwd}{Usa o atual diretório de trabalho?}
}
\value{

}
\description{
Sumariza conjuntos de dados NIR.
}
