% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nir_pkg.R
\name{write_NIRparams}
\alias{write_NIRparams}
\title{Write NIR parameters file to disk}
\usage{
write_NIRparams(file = "", wd = ".", surface = "", reads = "",
  nir_variables = "", subset_file = "", individual_id = "",
  individual_list = NULL, surface_id = "", group_id = "",
  nir_id = "X")
}
\arguments{
\item{file}{Dataset name.}

\item{wd}{Location of working directory. If not provided, current working directory will be used one.}

\item{surface}{Which leaf surface (abaxial, adaxial, both).}

\item{reads}{Do all the reads will be taken into account (all, mean)?}

\item{nir_variables}{Which NIR variables will be used (all, subset).}

\item{subset_file}{Location of NIR variable subset file, in case option selected in `nir_variables` was `subset`.}

\item{individual_id}{Which variable name corresponds to the individual?}

\item{individual_list}{Location of a file containing list of specimens to subset data when building datasets with `build_NIRdataset` function.}

\item{surface_id}{Which variable name corresponds to the leaf surface?}

\item{group_id}{Which variable name corresponds to the group category?}

\item{nir_id}{A string that can be used to grep column names containing NIR data. Default value is '`X`, which precedes all columns with NIR data.}
}
\value{
A message that indicates the file has been saved to disk.
}
\description{
Write NIR parameters file to disk
}
\examples{
# Write to disk a parameter file
write_NIRparams(file = "test", wd = getwd(), reads = "mean", surface = "abaxial", nir_variables = "all", surface_id = "face", individual_id = "especimenid", individual_list = NULL, group_id = "SP1", nir_id = "X")

# Read parameter file called "test" with function `read_NIRparams()`. It returns a data.frame
read_NIRparams("test-NIR.txt")

# You can also check parameter file with funciont `readLines()`
readLines("test-NIR.txt")
}
